Implement GetHTTP01Client in its own file
